requirements.txt
opencv-python
pyzmq
lz4
numpy